Improved image handling, now gets images from correct property. Pagination
Pagination works by fetching new posts until max of 10 shown is reached. To get a new post, dismiss one from current list. If last post is dismissed, the same one will be fetched every time. It's a bit quirky if items are dismissed quickly in a short time span, this could be improved.